+++ benches/bench_support.rs
@@ -1,6 +1,6 @@
 //! Provides support for benchmarks.
 
-use bench_diff::{DiffOut, statistics::AltHyp};
+use bench_diff::{DiffOut, LatencyUnit, bench_diff_with_status, statistics::AltHyp};
 use std::{
     hint::black_box,
     thread::{self},
@@ -10,7 +10,14 @@ use thread_map::{ThreadMap, ThreadMapLockError};
 const NTHREADS: i32 = 5;
 const NITER: i32 = 2_000;
 const FOLD_RATIO: i32 = 100;
-pub const EXEC_COUNT: usize = 1_000;
+const EXEC_COUNT: usize = 1_000;
+const NREPEATS: i32 = 100;
+
+fn print_params() {
+    println!(
+        "Params: NTHREADS={NTHREADS}, NITER={NITER}, FOLD_RATIO={FOLD_RATIO}, EXEC_COUNT={EXEC_COUNT}, NREPEATS={NREPEATS}"
+    );
+}
 
 fn update_value((i0, v0): &mut (i32, i32), j: i32) {
     *i0 = black_box(j);
@@ -23,7 +30,7 @@ pub trait Tm<V: Clone> {
     fn fold_values<W>(&self, z: W, f: impl Fn(W, &V) -> W) -> Result<W, ThreadMapLockError>;
 
     #[allow(unused)]
-    fn type_name(&self) -> &str;
+    fn type_name(&self) -> &'static str;
 }
 
 impl<V: Clone> Tm<V> for ThreadMap<V> {
@@ -39,12 +46,12 @@ impl<V: Clone> Tm<V> for ThreadMap<V> {
         self.fold_values(z, f)
     }
 
-    fn type_name(&self) -> &str {
+    fn type_name(&self) -> &'static str {
         "ThreadMap"
     }
 }
 
-pub fn tm_bench(tm: impl Tm<(i32, i32)> + Sync) {
+fn tm_bench(tm: impl Tm<(i32, i32)> + Sync) {
     let per_thread = |j: i32| {
         for i in 0..NITER {
             tm.with_mut(move |p: &mut (i32, i32)| update_value(p, j));
@@ -60,16 +67,16 @@ pub fn tm_bench(tm: impl Tm<(i32, i32)> + Sync) {
             }
         }
 
-        let sum = tm
-            .fold_values((0, 0), |acc, p| (acc.0 + p.0, acc.1 + p.1))
-            .unwrap();
+        // let sum = tm
+        //     .fold_values((0, 0), |acc, p| (acc.0 + p.0, acc.1 + p.1))
+        //     .unwrap();
         // println!(
         //     "type_name(tm)={}, thread_id={:?}, sum={:?}",
         //     tm.type_name(),
         //     thread::current().id(),
         //     sum
         // );
-        black_box(sum);
+        // black_box(sum);
     };
 
     thread::scope(|s| {
@@ -83,7 +90,7 @@ pub fn tm_bench(tm: impl Tm<(i32, i32)> + Sync) {
     });
 }
 
-pub fn print_diff_out(out: &DiffOut) {
+fn print_diff_out(out: &DiffOut) {
     const ALPHA: f64 = 0.05;
 
     println!();
@@ -102,4 +109,27 @@ pub fn print_diff_out(out: &DiffOut) {
     println!();
     println!("summary_f2={:?}", out.summary_f2());
     println!();
+}
+
+pub fn bench_compare<Tm1, Tm2>(ftm1: fn() -> Tm1, ftm2: fn() -> Tm2)
+where
+    Tm1: Tm<(i32, i32)> + Sync,
+    Tm2: Tm<(i32, i32)> + Sync,
+{
+    let f1 = || tm_bench(ftm1());
+    let f2 = || tm_bench(ftm2());
+
+    let type_name1 = ftm1().type_name();
+    let type_name2 = ftm2().type_name();
+
+    print_params();
+    println!();
+
+    for i in 1..NREPEATS {
+        eprintln!("*** i={i} ***");
+        let out = bench_diff_with_status(LatencyUnit::Nano, f1, f2, EXEC_COUNT, |_, _| {
+            println!(">>> bench_diff comparison: {type_name1} vs. {type_name2}; *** i={i} ***")
+        });
+        print_diff_out(&out);
+    }
 }